jwt auth added

+++ server/backend/app/views.py
@@ -4,14 +4,53 @@ from flask import request
 import pymysql
 import shortuuid
 from werkzeug.security import generate_password_hash, check_password_hash
+import jwt
+import datetime
+import time
 
 
+# Настройки БД
 db = pymysql.connect(host='f0762448.xsph.ru',
  user='f0762448_test_database',
  password='testpassword',
  database='f0762448_test_database',
  charset='utf8mb4',
  autocommit=True)
+
+
+# Создать токен аутентификации
+def encode_auth_token(user_id, login_time):
+    try:
+        payload = {
+            'exp': datetime.datetime.utcnow() + datetime.timedelta(days=30, seconds=0),
+            'iat': datetime.datetime.utcnow(),
+            'iss': 'ken',
+            'data': {
+                'id': user_id,
+                'login_time': login_time
+            }
+        }
+        return jwt.encode(
+            payload,
+            '123456',
+            algorithm='HS256'
+        )
+    except Exception as error:
+        return error
+
+# Проверить токен
+def decode_auth_token(auth_token):
+    try:
+        # Отменить подтверждение срока действия
+        payload = jwt.decode(auth_token, '123456', algorithms=['HS256'], options={'verify_exp': False})
+        if 'data' in payload and 'id' in payload['data']:
+            return payload
+        else:
+            raise jwt.InvalidTokenError
+    except jwt.ExpiredSignatureError:
+        return 'Token expired!'
+    except jwt.InvalidTokenError:
+        return 'Invalid token!'
 
 
 # выполнение sql запроса
@@ -62,6 +101,7 @@ def signup_user():
 def login_user():
     userid = None
     msg = None
+    token = None
     stat = None
     try:
         data = request.json
@@ -76,13 +116,15 @@ def login_user():
             raise Exception("Password is incorrect")
         else:
             userid = results[0][0]
-            msg = "Logged in!"
+            login_time = int(time.time())
+            token = encode_auth_token(userid, login_time)
+            msg = 'Logged in!'
             stat = 303
     except Exception as error:
         msg = str(error)
         stat = 500
     finally:
-        response = make_response(jsonify({'msg': msg, 'userid': userid}))
+        response = make_response(jsonify({'msg': msg, 'token': token, 'userid': userid}))
         response.status_code = stat
         response.mimetype = 'application/json'
     return response
@@ -95,18 +137,26 @@ def get_notes():
     msg = None
     stat = None
     try:
-        data = request.json
-        userid = data['userid']
+        if 'Authorization' in request.headers:
+            token = request.headers['Authorization']
+            payload = decode_auth_token(token)
 
-        query = "SELECT NoteID, Title, Body, Date, Edited FROM Notes WHERE UserID = '{}'".format(userid)
-        results = call_query(query)
+            if not isinstance(payload, str):
+                data = request.json
+                userid = data['userid']
 
-        if len(results) == 0:
-            raise Exception("There are no notes yet")
+                query = "SELECT NoteID, Title, Body, Date, Edited FROM Notes WHERE UserID = '{}'".format(userid)
+                results = call_query(query)
 
-        res = json.dumps(results)
-        msg = "Loaded"
-        stat = 200
+                if len(results) == 0:
+                    raise Exception("There are no notes yet")
+
+                res = json.dumps(results)
+                msg = "Loaded"
+                stat = 200
+            else:
+                msg = payload
+                stat = 403
     except Exception as error:
         msg = str(error)
         stat = 500
@@ -124,23 +174,26 @@ def update_note():
     msg = None
     stat = None
     try:
-        data = request.json
-        field = data['field']
-        value = data['value']
-        edited = data['edited']
-        noteid = data['noteid']
+        if 'Authorization' in request.headers:
+            token = request.headers['Authorization']
+            payload = decode_auth_token(token)
 
-        print(field)
+            if not isinstance(payload, str):
+                data = request.json
+                field = data['field']
+                value = data['value']
+                edited = data['edited']
+                noteid = data['noteid']
 
-        query = "UPDATE Notes SET {} = '{}', Edited = '{}' WHERE Notes.NoteID = '{}'".\
-            format(field, value, edited, noteid)
-        call_query(query)
+                query = "UPDATE Notes SET {} = '{}', Edited = '{}' WHERE Notes.NoteID = '{}'".\
+                    format(field, value, edited, noteid)
+                call_query(query)
 
-        # if len(results) == 0:
-        #     raise Exception("There are no notes yet")
-
-        msg = "Note updated"
-        stat = 200
+                msg = "Note updated"
+                stat = 200
+            else:
+                msg = payload
+                stat = 403
     except Exception as error:
         msg = str(error)
         stat = 500
@@ -159,19 +212,27 @@ def create_note():
     msg = None
     stat = None
     try:
-        data = request.json
-        title = data['title']
-        value = data['value']
-        date = data['date']
-        userid = data['userid']
-        noteid = shortuuid.ShortUUID().random(length=16)
+        if 'Authorization' in request.headers:
+            token = request.headers['Authorization']
+            payload = decode_auth_token(token)
 
-        query = "INSERT INTO Notes (NoteID, UserID, Body, Title, Date, Edited) " \
-                "VALUES ('{}', '{}', '{}', '{}', '{}', '{}')".format(noteid, userid, value, title, date, date)
-        call_query(query)
+            if not isinstance(payload, str):
+                data = request.json
+                title = data['title']
+                value = data['value']
+                date = data['date']
+                userid = data['userid']
+                noteid = shortuuid.ShortUUID().random(length=16)
 
-        msg = "Note created"
-        stat = 201
+                query = "INSERT INTO Notes (NoteID, UserID, Body, Title, Date, Edited) " \
+                        "VALUES ('{}', '{}', '{}', '{}', '{}', '{}')".format(noteid, userid, value, title, date, date)
+                call_query(query)
+
+                msg = "Note created"
+                stat = 201
+            else:
+                msg = payload
+                stat = 403
     except Exception as error:
         msg = str(error)
         stat = 500
@@ -189,20 +250,28 @@ def delete_note():
     stat = None
     query = ''
     try:
-        data = request.json
-        noteids = tuple(data['noteids'])
+        if 'Authorization' in request.headers:
+            token = request.headers['Authorization']
+            payload = decode_auth_token(token)
 
-        if len(noteids) > 1:
-            query = "DELETE FROM Notes WHERE NoteID IN {}".format(noteids)
-        elif len(noteids) == 1:
-            query = "DELETE FROM Notes WHERE NoteID = '{}'".format(noteids[0])
-        else:
-            raise Exception("At least one note should be selected")
+            if not isinstance(payload, str):
+                data = request.json
+                noteids = tuple(data['noteids'])
 
-        call_query(query)
+                if len(noteids) > 1:
+                    query = "DELETE FROM Notes WHERE NoteID IN {}".format(noteids)
+                elif len(noteids) == 1:
+                    query = "DELETE FROM Notes WHERE NoteID = '{}'".format(noteids[0])
+                else:
+                    raise Exception("At least one note should be selected")
 
-        msg = "Note deleted"
-        stat = 200
+                call_query(query)
+
+                msg = "Note deleted"
+                stat = 200
+            else:
+                msg = payload
+                stat = 403
     except Exception as error:
         msg = str(error)
         stat = 500
